Moved PlanningHeader files closer to PlanningOverview where it is used
